Break Tests class apart
