+++ src/Verify.MSTest.SourceGenerator.Tests/GlobalNamespaceTests.HasAttributeOnClass.verified.txt
