+++ src/Verify.MSTest.SourceGenerator.Tests/GlobalNamespaceTests.cs
@@ -1,0 +1,17 @@
+﻿public class GlobalNamespaceTests(ITestOutputHelper output) : TestBase(output)
+{
+    [Fact]
+    public Task HasAttributeOnClass()
+    {
+        var source = """
+            using VerifyMSTest;
+
+            [UsesVerify]
+            public partial class Foo
+            {
+            }
+            """;
+
+        return VerifyGenerator(TestDriver.Run(source));
+    }
+}
