+++ src/Verify.MSTest.SourceGenerator.Tests/InheritanceTests.cs
@@ -1,0 +1,22 @@
+public class InheritanceTests(ITestOutputHelper output) : TestBase(output)
+{
+    [Fact]
+    public Task HasAttributeOnBaseAndDerivedClasses()
+    {
+        var source = """
+            using VerifyMSTest;
+
+            [UsesVerify]
+            public partial class Base
+            {
+            }
+
+            [UsesVerify]
+            public partial class Derived : Base
+            {
+            }
+            """;
+
+        return VerifyGenerator(TestDriver.Run(source));
+    }
+}
