+++ src/Verify.MSTest.SourceGenerator.Tests/NamespaceTests.HasAttributeOnClass.verified.txt
