+++ src/Verify.MSTest.SourceGenerator.Tests/NamespaceTests.HasAttributeOnClassWithGenericsInNestedNamespace.verified.txt
